requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
structlog==25.4.0

streamlit==1.49.1

sentence-transformers==5.1.0
faiss-cpu==1.12.0
openai==1.100.2
pypdf==6.0.0
python-docx==1.2.0

pytest==8.4.1
httpx==0.28.1
